學習分析hw3/requirements.txt
line-bot-sdk
flask
configparser
azure.ai.textanalytics